requirements.txt
transformers
datasets
evaluate
accelerate
scikit-learn
tensorboard
pandas
jupyterlab
ipywidgets
ablang
fair-esm
rjieba
sentencepiece
annotated-types
biopython
fire
polars
pydantic
pydantic-core
termcolor
typing-extensions